feat[step-1]: refactoring
