+++ jest.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  preset: 'react-native',
-};
+    preset: 'react-native',
+  };
