+++ src/App.tsx
@@ -2,6 +2,6 @@ import { SafeAreaView } from 'react-native';
 import { Home } from './components/home/Home';
 
 export const App: React.FC = () => 
-    <SafeAreaView>
-        <Home />
-    </SafeAreaView>
+  <SafeAreaView>
+    <Home />
+  </SafeAreaView>
