+++ src/components/home/components/article-list/ArticleList.tsx
@@ -4,9 +4,9 @@ import { Article } from '../article/Article';
 import { ARTICLES_DATA } from './articles-data';
 
 export const ArticleList: React.FC = () => 
-    <FlatList
-      className="container"
-      data={ARTICLES_DATA}
-      renderItem={({item}) => <Article data={item} />}
-      keyExtractor={item => item.id}
-    />
+  <FlatList
+    className="container"
+    data={ARTICLES_DATA}
+    renderItem={({item}) => <Article data={item} />}
+    keyExtractor={item => item.id}
+   />
