+++ src/components/home/components/article-list/articles-data.ts
@@ -1,0 +1,39 @@
+export const ARTICLES_DATA = [
+  {
+    id: '1',
+    category: 'Finance',
+    date: 'July 4, 2022, 10:05',
+    title: 'Is Bitcoin the new gold?',
+    content:
+      'There has been talk that bitcoin is the new gold. The esteemed Satoshi gave his opinion on the matter.',
+    pictureUrl: require('../../assets/articles-images/1.png'),
+  },
+  {
+    id: '2',
+    category: 'World',
+    date: 'July 4, 2022, 09:42',
+    title: 'Five Boroughs, Five Days, One Question',
+    content:
+      'As millions of New Yorkers entered a third pandemic summer, we crisscrossed the city to find out: What’s the vibe right now?',
+    pictureUrl: require('../../assets/articles-images/2.png'),
+  },
+  {
+    id: '3',
+    category: 'Automobiles',
+    date: 'July 3, 2022, 19:34',
+    title: 'Mercedes-Benz recalls nearly 1 million vehicles over brake issue.',
+    content:
+      'The German carmaker said possible corrosion could cause the brakes in two of its S.U.V.s and in a luxury minivan model to malfunction.',
+    pictureUrl: require('../../assets/articles-images/3.png'),
+  },
+  {
+    id: '4',
+    category: 'Real Estate',
+    date: 'July 3, 2022, 17:02',
+    title: 'Two Homes That Nourish a Family’s Many Creative Pursuits',
+    content:
+      'A businesswoman with a husband, three children, three restaurants and a publishing venture has a home upstate and a studio in Manhattan to make it all work.',
+    pictureUrl: require('../../assets/articles-images/4.png'),
+  },
+];
+
